requirements.txt

djangorestframework
mysqlclient
